+++ Assets/uNode3/Core/Nodes/Event/StateOnEnterEvent.cs
@@ -3,19 +3,19 @@ using UnityEngine;
 using System.Collections.Generic;
 
 namespace MaxyGames.UNode.Nodes {
-    [EventMenu("", "On Enter", scope = NodeScope.State)]
+    [EventMenu("", "On State Enter", scope = NodeScope.State)]
 	[StateEvent]
 	[Description("On Enter is called once when the state becomes active.")]
 	public class StateOnEnterEvent : BaseGraphEvent {
 		public override void OnRuntimeInitialize(GraphInstance instance) {
 			if(nodeObject.parent is NodeObject parentNode && parentNode.node is INodeWithEnterExitEvent state) {
-				state.OnEnterState += (flow) => {
+				state.OnEnterCallback += (flow) => {
 					Trigger(flow);
 				};
 			}
 		}
 
-		public override string GetTitle() => "On Enter";
+		public override string GetTitle() => "On State Enter";
 
 		public override Type GetNodeIcon() {
 			return typeof(TypeIcons.EventIcon);