+++ Assets/uNode3/Core/Nodes/StateNode.cs
@@ -25,7 +25,7 @@ namespace MaxyGames.UNode.Nodes {
 		}
 
 		private event System.Action<Flow> m_onEnter;
-		public event System.Action<Flow> OnEnterState {
+		public event System.Action<Flow> OnEnterCallback {
 			add {
 				m_onEnter -= value;
 				m_onEnter += value;
@@ -35,7 +35,7 @@ namespace MaxyGames.UNode.Nodes {
 			}
 		}
 		private event System.Action<Flow> m_onExit;
-		public event System.Action<Flow> OnExitState {
+		public event System.Action<Flow> OnExitCallback {
 			add {
 				m_onExit -= value;
 				m_onExit += value;