+++ Assets/uNode3/Core/StateMachines/StateNodes/StateTransition.cs
@@ -33,7 +33,7 @@ namespace MaxyGames.UNode.Nodes {
 		}
 
 		private event System.Action<Flow> m_onEnter;
-		public event System.Action<Flow> OnEnterState {
+		public event System.Action<Flow> OnEnterCallback {
 			add {
 				m_onEnter -= value;
 				m_onEnter += value;
@@ -43,7 +43,7 @@ namespace MaxyGames.UNode.Nodes {
 			}
 		}
 		private event System.Action<Flow> m_onExit;
-		public event System.Action<Flow> OnExitState {
+		public event System.Action<Flow> OnExitCallback {
 			add {
 				m_onExit -= value;
 				m_onExit += value;