+++ Assets/uNode3/Core.Editor/Classes/CustomInspector.cs
@@ -200,13 +200,7 @@ namespace MaxyGames.UNode.Editors {
 		public static void Inspect(Vector2 position, GraphEditorData editorData, int limitMultiEdit = 5) {
 			ActionPopupWindow.Show(() => {
 				ShowInspector(editorData);
-			}).ChangePosition(position).autoFocus = false;
-		}
-
-		public static void Inspect(Vector2 position, object value, UnityEngine.Object unityObject = null, BindingFlags flags = BindingFlags.Public | BindingFlags.Instance) {
-			ActionPopupWindow.Show(() => {
-				uNodeGUIUtility.ShowFields(value, unityObject, flags);
-			}).ChangePosition(position).autoFocus = false;
+			}).ChangePosition(position);
 		}
 
 		private static void RenameObject(UnityEngine.Object obj, Rect rect) {